+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### v0.7.1
 
+* Relaxed checks made on `X` in `_validate_and_reformat_input()` since that
+  is the concern of the underlying estimator and not Fairlearn
 
 
 ### v0.7.0